+++ client/views/users/user_item.js
@@ -71,5 +71,10 @@ Template.user_item.events({
 				isAdmin: false
 			}
 		});
+	},
+	'click .delete-link': function(e, instance){
+		e.preventDefault();
+		if(confirm("Are you sure you want to delete "+getDisplayName(instance.data)+"?"))
+			Meteor.users.remove(instance.data._id);
 	}
 })